+++ addOns/pscanrules/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - The Private Address Disclosure and Session ID in URL Rewrite scan rules now include example alert functionality for documentation generation purposes (Issue 6119 and 7100).
 
+### Fixed
+- The Modern App Detection scan rule now ignores non-HTML files (Issue 7617).
+
 ## [44] - 2022-10-27
 ### Added
 - The following scan rules were added, having been promoted from Beta:
@@ -28,7 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - User Controlled Javascript Event
   - User Controlled Open Redirect
   - X-Backend-Server Information Leak
-  - X-ChromeLogger-Data Info Leak   
+  - X-ChromeLogger-Data Info Leak
 
 ### Changed
 - Update minimum ZAP version to 2.12.0.
@@ -234,7 +237,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Maintenance changes.
 - Migrate CSP Scanner into the main passive scan release package (promoting it to Release). Upgrade Salvation (dependency) to 2.6.0.
-- Application Error scanner change for HTTP 500. Alert changed to low risk for HTTP 500, and not raised at all when Threshold is High.  
+- Application Error scanner change for HTTP 500. Alert changed to low risk for HTTP 500, and not raised at all when Threshold is High.
 - Updated the reference link for the alert: Web Browser XSS Protection Not Enabled.
 - Promote Charset Mismatch Scanner to release (Issue 4460).
 - Promote ViewState Scanner to release (Issue 4453).